Reintegrate objective function, clean up for local running
